# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 'FlixFlex API Docs'
  description: ''
  version: 1.0.0
servers:
  -
    url: 'http://localhost:8000'
paths:
...
  /api/movies:
    get:
      summary: 'Get All Movies'
      operationId: getAllMovies
      description: 'This endpoint lets you get all movies in batches of 10.'
      parameters:
        -
          in: query
          name: page
          description: 'Page you wanna get, Defaults to 1.'
          example: 18
          required: false
          schema:
            type: integer
            description: 'Page you wanna get, Defaults to 1.'
            example: 18
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Movies and Shows Routes'
  /api/tv-series:
    get:
      summary: 'Get All Tv Series'
      operationId: getAllTvSeries
      description: 'This endpoint lets you get all tv series in batches of 10.'
      parameters:
        -
          in: query
          name: page
          description: 'Page you wanna get, Defaults to 1.'
          example: 5
          required: false
          schema:
            type: integer
            description: 'Page you wanna get, Defaults to 1.'
            example: 5
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Movies and Shows Routes'
  /api/popular-movies:
    get:
      summary: 'Get Popular Movies'
      operationId: getPopularMovies
      description: 'This endpoint lets you get top 5 rated movies.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Movies and Shows Routes'
  /api/popular-tv-series:
    get:
      summary: 'Get Popular Tv Series'
      operationId: getPopularTvSeries
      description: 'This endpoint lets you get top 5 rated tv series.'
      parameters:
        -
          in: header
          name: Authorization
          description: ''
          example: 'Bearer {YOUR_AUTH_KEY}'
          schema:
            type: string
        -
          in: header
          name: Content-Type
          description: ''
          example: application/json
          schema:
            type: string
        -
          in: header
          name: Accept
          description: ''
          example: application/json
          schema:
            type: string
      responses:
        401:
          description: ''
          content:
            application/json:
              schema:
                type: object
                example:
                  message: Unauthenticated.
                properties:
                  message:
                    type: string
                    example: Unauthenticated.
      tags:
        - 'Movies and Shows Routes'
...
components:
  securitySchemes:
    default:
      type: http
      scheme: bearer
      description: ''
security:
  -
    default: []